requirements.txt
streamlit>=1.38.0 
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-groq==0.2.5
python-dotenv>=1.0.1